feat(gh-99): use pixel for standard space props (#8)
* refactor(GH-99): let projects define space props

but we can give sample standards

* refactor(GH-99): set standard space props

Because only core-styles uses them; later, usage could move back to CMS, but no reason to do so now. If stuff is sharable, let it be shared, until we learn it is specific, because sharing is already hard at TACC, so reduce barrier to share.