fix(components): django-cms-forms pattern (cms)
This is for CMS, but was being tested here. it works now.

Latest real bug was misssing the .html extension inthe demo.

This pattern was just committed in TACC/Core-CMS:
- task/get-core-styles-beyond-0.11.0
- d13ca50

I am porting the polishes and fixes here just so that when I delete I am deleting at the time at which CMS resumes control of it.