fix(components): no card pattern in modals
A follow-up to previous commit. The Portal modals (any modals) are not within the #root, nor even <main>. So I am disabling TACC cards inside modals.

Ideally, TACC/tup-ui doesn't port Bootstrap (Reactstrap), but I do not have bandwidth to fight that battle right now. THis is an easier fix that is just as easy to remove.