fix: restore s-paragraph-table (removed in 0a9b0b6)
This table is a Trump (Scope) pattern used by CMS.

It is loaded by table to keep code DRY, but it should appear in Trumps section of demo.

I welcome discussion of different solutions. This is not the only such case like this.

P.S. The s-truncated-table was removed in 0a9b0b6 also. That's expected. No clients use it. If they want truncation, they can use Core-Styles truncation or do it themselves.